backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
axios==0.4.0
black==25.1.0
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
openai==1.97.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
requests-toolbelt==1.0.0
RG==0.0.75
rich==14.1.0
s3transfer==0.13.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
gunicorn==22.0.0
whitenoise==6.7.0
